Rename files to be less verbose
