+++ static/app/components/searchQueryBuilder/index.stories.tsx
@@ -2,7 +2,7 @@ import {Fragment} from 'react';
 import styled from '@emotion/styled';
 
 import Alert from 'sentry/components/alert';
-import {SearchQueryBuilder} from 'sentry/components/searchQueryBuilder/searchQueryBuilder';
+import {SearchQueryBuilder} from 'sentry/components/searchQueryBuilder';
 import SizingWindow from 'sentry/components/stories/sizingWindow';
 import storyBook from 'sentry/stories/storyBook';
 import type {TagCollection} from 'sentry/types';
